moved vehicle function to own handler.
